Create Complex.java
This is a Java class that implements complex numbers (numbers that have a real and an imaginary part). It has a variety of methods from adding up two complex numbers to getting the n-th root of a complex number by first presenting it in polar form (instead of a cartesian coordinate system).